+++ src/components/AboutCard/aboutCard.module.scss
