+++ src/components/FormRegistration/formRegistration.module.scss
