+++ src/components/LoveIcon.js
