+++ src/components/Place/place.module.scss
